backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.36
alembic==1.13.3
psycopg2-binary==2.9.9
supabase==2.10.0

# Cache / Queue
redis==5.2.0
celery==5.4.0

# HTTP client (for calling external APIs)
httpx==0.27.2
aiohttp==3.10.10

# AI / ML
sentence-transformers==3.3.1
faiss-cpu==1.9.0
lightgbm==4.5.0
xgboost==2.1.3
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3

# HuggingFace sentiment model
transformers==4.46.3
torch==2.5.1

# Claude API
anthropic==0.40.0

# Reddit
praw==7.8.1

# ML tracking
mlflow==2.18.0

# Monitoring
prometheus-client==0.21.0

# Utilities
python-dotenv==1.0.1
pydantic==2.10.1
pydantic-settings==2.6.1
